HTTPCookie is a subtype of HTTPClient to not collide with Foundations top-level type
