+++ Tests/NIOHTTPClientTests/HTTPCookieTests.swift
@@ -19,7 +19,7 @@ import XCTest
 class HTTPCookieTests: XCTestCase {
     func testCookie() {
         let v = "key=value; Path=/path; Domain=example.com; Expires=Wed, 21 Oct 2015 07:28:00 GMT; Max-Age=42; Secure; HttpOnly"
-        let c = HTTPCookie(from: v, defaultDomain: "exampe.org")!
+        let c = HTTPClient.HTTPCookie(from: v, defaultDomain: "exampe.org")!
         XCTAssertEqual("key", c.name)
         XCTAssertEqual("value", c.value)
         XCTAssertEqual("/path", c.path)
@@ -32,7 +32,7 @@ class HTTPCookieTests: XCTestCase {
 
     func testCookieDefaults() {
         let v = "key=value"
-        let c = HTTPCookie(from: v, defaultDomain: "example.org")!
+        let c = HTTPClient.HTTPCookie(from: v, defaultDomain: "example.org")!
         XCTAssertEqual("key", c.name)
         XCTAssertEqual("value", c.value)
         XCTAssertEqual("/", c.path)
